Broadcast library for GERTi, requested by MGR
GERTiExtended is a library built to expand the functionality of the GERT system, enabling multicasts and broadcasts to be sent
across multiple networks if a connection is able to be made. Current functions of this library are 
GERTiExtended.setNet() which takes a table of addresses and creates sockets for each.
GERTiExtended.Netcast() which takes data, and a selective list of addresses to broadcast to, if left empty this will default to all
GERTiExtended.getStreams() This command gets a list of all currently connected broadcast streams.
GERTiExtended.flush() will destroy all streams in the table, closing the connection on both sides.