GERTInstaller
GERTInstaller is an automated installer for the GERTi system. It can automatically install GERTi clients and GERTi servers.
Combined with the simple foxlib UI the system install is quick and painless. If any issues arise please report to Nick on the Global Empire server listed in the readme file of the main directory of this repo.