Release 2! Introducing Encryption (#35)
* OpenFTP

A program that allows FTP transfers.

* Reordering of files + work on Version 2

* Update FTPServer - GERTi.lua

* Update README.txt

Included secure OS recommendation

* Release 2

This is the finished Release 2 of the OpenFTP server, which supports T3 Data Cards for fully encrypted data communication between client and server. It does not support automatic data card detection, and as such Release 1 might be preferable for some.