Emergency patch (#48)
* Create FTPServer - GERTi.lua

* Create FTPClient - GERTi-library mode.lua

* Update FTPClient - GERTi-library mode.lua

* Update FTPClient - GERTi-library mode.lua

* ls feature

Shoehorned in an LS feature. Use it by requesting a file from the server with / at the end of the filepath. The client function will return 3 parameters: false, a string, and a serialised table of all elements in the requested directory. subdirectories have a slash at the end.

* behavioural changes

* Functional

This took me too long